+++ backend/sokdak/src/test/java/com/wooteco/sokdak/notification/handler/DeletionEventHandlerTest.java
@@ -1,12 +1,16 @@
 package com.wooteco.sokdak.notification.handler;
 
+import static com.wooteco.sokdak.notification.domain.NotificationType.COMMENT_REPORT;
+import static com.wooteco.sokdak.notification.domain.NotificationType.HOT_BOARD;
+import static com.wooteco.sokdak.notification.domain.NotificationType.NEW_COMMENT;
+import static com.wooteco.sokdak.notification.domain.NotificationType.NEW_REPLY;
 import static org.assertj.core.api.Assertions.assertThat;
 
-import com.wooteco.sokdak.comment.event.CommentDeletionEvent;
+import com.wooteco.sokdak.comment.domain.CommentDeletionEvent;
 import com.wooteco.sokdak.config.AsyncTestConfig;
 import com.wooteco.sokdak.notification.domain.Notification;
 import com.wooteco.sokdak.notification.repository.NotificationRepository;
-import com.wooteco.sokdak.post.event.PostDeletionEvent;
+import com.wooteco.sokdak.post.domain.PostDeletionEvent;
 import com.wooteco.sokdak.util.DatabaseCleaner;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
@@ -37,8 +41,8 @@ class DeletionEventHandlerTest {
     @Test
     void handleCommentDeletion() {
         Long commentId = 1L;
-        notificationRepository.save(Notification.commentReport(1L, 1L, commentId));
-        notificationRepository.save(Notification.newReply(1L, 1L, commentId));
+        notificationRepository.save(new Notification(COMMENT_REPORT, 1L, 1L, commentId));
+        notificationRepository.save(new Notification(NEW_REPLY, 1L, 1L, commentId));
         CommentDeletionEvent commentDeletionEvent = new CommentDeletionEvent(commentId);
 
         deletionEventHandler.handleCommentDeletion(commentDeletionEvent);
@@ -50,8 +54,8 @@ class DeletionEventHandlerTest {
     @Test
     void handlePostDeletion() {
         Long postId = 1L;
-        notificationRepository.save(Notification.postHotBoard(1L, postId));
-        notificationRepository.save(Notification.newComment(1L, postId));
+        notificationRepository.save(new Notification(HOT_BOARD, 1L, postId, null));
+        notificationRepository.save(new Notification(NEW_COMMENT, 1L, postId, null));
         PostDeletionEvent postDeletionEvent = new PostDeletionEvent(postId);
 
         deletionEventHandler.handlePostDeletion(postDeletionEvent);
